+++ isotopes.c
@@ -900,12 +900,12 @@ punch_isotopes(void)
 		if (punch.high_precision == FALSE)
 		{
 			fpunchf(sformatf("I_%s", punch.isotopes[i].name), "%12.4e\t",
-					iso);
+				(double) iso);
 		}
 		else
 		{
 			fpunchf(sformatf("I_%s", punch.isotopes[i].name), "%20.12e\t",
-					iso);
+				(double) iso);
 		}
 
 	}
@@ -940,12 +940,12 @@ punch_calculate_values(void)
 		if (punch.high_precision == FALSE)
 		{
 			fpunchf(sformatf("V_%s", punch.calculate_values[i].name),
-					"%12.4e\t", result);
+				"%12.4e\t", (double) result);
 		}
 		else
 		{
 			fpunchf(sformatf("V_%s", punch.calculate_values[i].name),
-					"%20.12e\t", result);
+				"%20.12e\t", (double) result);
 		}
 	}
 	return (OK);
